server/requirements.txt
django
djangorestframework
django-cors-headers
numpy
firebase-admin
tensorflow==2.15.0
keras==2.15.0
keras_applications==1.0.8
mysql-connector-python
scikit-learn==1.2.2
transformers
kapre
mysqlclient
imblearn
pandas
librosa
matplotlib
sounddevice
pyaudio
pydub
pyjwt